refactor(command_processor): remove dead exports and duplicate ServiceCommandBuilder
Delete plugins' ServiceCommandBuilder class (~650 lines) which had only 1 method
used in production. Move create_certificate_generation_command to CommandUtils.
Trim 13 unused re-exports from command_processor/__init__.py.
