+++ panther/plugins/services/base/__init__.py
@@ -3,11 +3,9 @@
 from .http_service_base import BaseHTTPServiceManager
 from .minip_service_base import BaseMinipServiceManager
 from .quic_service_base import BaseQUICServiceManager
-from .service_command_builder import ServiceCommandBuilder
 
 __all__ = [
     "BaseQUICServiceManager",
     "BaseHTTPServiceManager",
     "BaseMinipServiceManager",
-    "ServiceCommandBuilder",
 ]
